Fixed EBNF to indicate precedence between alternation and concatenation
The grammar provided by Wikipedia doesn’t properly represent the concatenation operator as binding more tightly than the alternation operator. It’s not clear why that detail was omitted given that it wouldn’t be hard to do in EBNF, or even BNF. Regardless, this commit fixes the problem, albeit rather simplistically.